fk panens, lahan, dan tanaman

+++ app/Controllers/Tanaman.php
@@ -97,16 +97,16 @@ class Tanaman extends BaseController
             return $this->errorJson('Tidak ditemukan', 404);
         }
 
-        // $used = db_connect()
-        //     ->table('panen')
-        //     ->where('tanaman_id', $id)
-        //     ->countAllResults();
+        $used = db_connect()
+            ->table('panen')
+            ->where('tanaman_id', $id)
+            ->countAllResults();
 
-        // if ($used > 0) {
-        //     return $this->errorJson(
-        //         'Tanaman tidak dapat dihapus karena sudah digunakan dalam pencatatan panen.'
-        //     );
-        // }
+        if ($used > 0) {
+            return $this->errorJson(
+                'Tanaman tidak dapat dihapus karena sudah digunakan dalam pencatatan panen.'
+            );
+        }
 
         $this->model->delete($id);
